requirements.txt
# Python dependencies
python-telegram-bot[job-queue]>=21.0